[Build][Android] Enable release build and use same versioning scheme
We need to start distributing CameraKit sample apps to AppCenter therefore we need to generate release builds that are properly versioned. This PR enables Proguarded release build as well as updates the Android build scripts to use the root VERSION file to locate dependencies and to use as `versionName` of the release apk build. Also, we base the `versionCode` on git's ref count so that we don't need to manually increase the version code - this gets removed from the build script once the app gets ejected.
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/19
Cooled by jokubas-dargis