[Build] Create a script and CI job to publish ejected distribution to Github
We are preparing to release CameraKit publicly and we are going to have a Github repository under https://github.com/Snapchat where we will mirror the CameraKit samples (and some docs). This PR creates a script that runs a build on the current branch, gets ejected distribution as a directory and copies it over to the Github repo clone. The files copied into the repo get pushed into a unique branch on Github repo and then a draft PR is opened ready to be merged.

The CI job is setup here: https://snapengine-builder.sc-corp.net/jenkins/job/camerakit-distribution-publish-github/build?delay=0sec - it allows to specify the Github repo we want to publish to as well as username and auth token of the user that will communicate with Github. Currently, I created a personal access token for my `eleventigerssc` that I used in the test builds however we might need to create a bot user on Github to perform the publishing if needed.

Here's the example PR that gets created: https://github.com/Snap-Kit/camera-kit-samples-test/pull/2
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/463
Cooled by jokubas-dargis