[Doc] Update RELEASING doc with more up-to-date steps
Current RELEASING doc is slightly out-of-date due to changes in our CI flows etc. we need to update and also link to it from the main README to make it clear that it should be used.
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/485
Cooled by jokubas-dargis