[Doc][Build] Fix generated CHANGELOG
Since 1.6.12 the releases were not properly tagged and entries were manually added to the CHANGELOG in both release/1.6.0 and release/1.7.0 branches. This lead to a very difficult to repair git history which resulted in generated CHANGELOG that has various entries messed up, interleaved etc. I tried to manually rebase the release/1.6.0 with some success however when the rebased history was combined with the release/1.7.0 branch to generate the CHANGELOG the `git-chglog` tool we use couldn't properly order the commits/tags resulting in interleaved/missing entries.

The only possible solution to fix this now is to consider releases <= 1.7.0 tainted and not generate the CHANGELOG from that history - we can append this partially manual partially generated CHANGELOG to a new, fresh CHANGELOG that starts with 1.7.1.

This PR updates the `git-chglog` binary that we use in the script so that we can use the "semver" entry sorting. Our script is also updated to append an old CHANGELOG to the newly generated one starting with 1.7.1. The old CHANGELOG is committed in with `.old.md` suffix while the main CHANGELOG.md gets updated to a clean version. 
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/534
Cooled by jokubas-dargis