[IOS] CameraKitSample - Assert API Token is specified
**Background**
If a 3rd party developer tries and runs ios/CameraKitSample immediately without specifying an API token, they would encounter a situation where "REPLACE-THIS-WITH-YOUR-OWN-APP-SPECIFIC-VALUE" is saved in UserDefaults as current API token value. This causes the demo app to crash continuously due to an invalid token / state, and the only way to recover is to delete / reinstall the app.

**Change**
This fix prevents the situation described above by crashing the app *before* saving the token in UserDefaults if the client token has not been specified and equals "REPLACE-THIS-WITH-YOUR-OWN-APP-SPECIFIC-VALUE".
---
Automatic squash commit from https://github.sc-corp.net/Snapchat/camera-kit-distribution/pull/1308
Cooled by psharanda